requirements.txt
streamlit
yfinance
google-generativeai
plotly
pandas